rename organization controller file
